backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
langchain==0.2.17
langchain-community==0.2.17
langchain-groq==0.1.9
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
faiss-cpu==1.8.0
pypdf==4.2.0
sentence-transformers==3.0.1
pydantic==2.7.4
httpx==0.27.0
huggingface-hub==0.24.7
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
redis==5.0.4
psycopg2-binary==2.9.9
langgraph==0.1.19